requirements.txt
streamlit==1.28.2
langchain==0.1.0
langchain-community==0.0.32
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.2.2
PyGithub==2.1.1
tree-sitter==0.20.4
tree-sitter-languages==1.8.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
gitpython==3.1.40
pygments==2.17.2
markdown==3.5.1
beautifulsoup4==4.12.2
ragas==0.1.0